Move Keras integration into its own module
