+++ verta/docs/reference/api/integrations.rst
@@ -1,5 +1,5 @@
 Integrations
 ============
 
-.. automodule:: verta.integrations
+.. automodule:: verta.integrations.keras
     :members:
